refactor(operator): colocate observed actuator
Keep the observed actuator alongside the package's other actuator implementation and remove its redundant unit tests.
